Adds str_path handling to get_item and get_section which means that this closes #118 -- you can have the path parsed implicitly, no need to parse it beforehand
No more str_path_separator overrides accepted -- it was too expensive and complicated to pass around overrides and then to calculate what needs to be used in each case. User needs to be strict. If they need to load data from a dirty source, they need to clean it up themselves. Replacing separators in the input dictionary isn't hard.